cleanup posting matches
Should now be able to call matchPusher with a provided directory that contains matchJsons. That will send batches of 5000 matches (10000 rows) to the web site. I did run into an issue at the end of processing locally but I don't think it impacted the number of rows pushed? There is a try catch so I don't think it stopped any matches, but something to keep an eye on.